examen

+++ src/main/java/com/perfulandia/perfulandia/repository/PerfumeRepository.java
@@ -2,14 +2,12 @@ package com.perfulandia.perfulandia.repository;
 
 import com.perfulandia.perfulandia.model.Perfume;
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
-import java.util.ArrayList;
 
 @Repository
-public interface PerfumeRepository extends JpaRepository<Perfume, Long> {
+public interface PerfumeRepository extends JpaRepository<Perfume, Integer> {
 
     Boolean existsById(int id);
     Perfume findById(int id);
